ironclad manager bugfixes
