+++ api-module-library/ironclad/package.json
@@ -1,6 +1,6 @@
 {
     "name": "@friggframework/api-module-ironclad",
-    "version": "0.0.2",
+    "version": "0.0.4",
     "prettier": "@friggframework/prettier-config",
     "description": "",
     "main": "index.js",
