--- api-module-library/slack/CHANGELOG.md
@@ -1,4 +1,0 @@
-# v0.0.1 (Sep 27 2022)
-
-#### Generated
-- Initialized from template
