--- api-module-library/slack/README.md
@@ -1,3 +1,0 @@
-# Frigg API Module Slack
-
-A quick description of Slack.
