--- api-module-library/slack/defaultConfig.json
@@ -1,9 +1,0 @@
-{
-    "name": "slack",
-    "label": "slack",
-    "productUrl": "",
-    "apiDocs": "",
-    "logoUrl": "https://friggframework.org/assets/img/slack.jpeg",
-    "categories": [],
-    "description": "slack"
-}
