--- api-module-library/slack/index.js
@@ -1,13 +1,0 @@
-const { Api } = require('./api');
-const { Credential } = require('./models/credential');
-const { Entity } = require('./models/entity');
-const ModuleManager = require('./manager');
-const Config = require('./defaultConfig');
-
-module.exports = {
-  Api,
-  Credential,
-  Entity,
-  ModuleManager,
-  Config,
-};
