--- api-module-library/slack/jest-setup.js
@@ -1,2 +1,0 @@
-const { globalSetup } = require('@friggframework/test-environment');
-module.exports = globalSetup;
