--- api-module-library/slack/jest-teardown.js
@@ -1,2 +1,0 @@
-const { globalTeardown } = require('@friggframework/test-environment');
-module.exports = globalTeardown;
