--- api-module-library/slack/models/entity.js
@@ -1,10 +1,0 @@
-'use strict';
-
-const mongoose = require('mongoose');
-const { Entity: Parent } = require('@friggframework/module-plugin');
-
-const schema = new mongoose.Schema({});
-const name = 'SlackEntity';
-const Entity = Parent.discriminators?.[name] || Parent.discriminator(name, schema);
-
-module.exports = { Entity };
