--- api-module-library/slack/package.json
@@ -1,26 +1,0 @@
-{
-	"name": "@friggframework/api-module-slack",
-	"version": "0.1.0",
-	"prettier": "@friggframework/prettier-config",
-	"description": "",
-	"main": "index.js",
-	"scripts": {
-			"lint:fix": "prettier --write --loglevel error . && eslint . --fix",
-			"test": "jest"
-	},
-	"author": "",
-	"license": "MIT",
-	"devDependencies": {
-		"@friggframework/eslint-config": "^1.0.7",
-		"@friggframework/test-environment": "^1.1.4",
-		"eslint": "^8.22.0",
-		"jest": "^28.1.3",
-		"prettier": "^2.7.1",
-		"sinon": "^14.0.0"
-	},
-	"dependencies": {
-		"@friggframework/assertions": "^1.0.4",
-		"@friggframework/module-plugin": "^1.0.12"
-	},
-	"gitHead": "1ff6f32fa238512761a9c361f0e07f49b6bf5ed7"
-}
