--- api-module-library/slack/test/manager.test.js
@@ -1,27 +1,0 @@
-const Manager = require('../manager');
-const mongoose = require('mongoose');
-const config = require('../defaultConfig.json');
-require('dotenv').config();
-
-describe(`Should fully test the ${config.label} Manager`, () => {
-    let manager, userManager;
-
-    beforeAll(async () => {
-        await mongoose.connect(process.env.MONGO_URI);
-        manager = await Manager.getInstance({
-            userId: new mongoose.Types.ObjectId(),
-        });
-    });
-
-    afterAll(async () => {
-        await Manager.Credential.deleteMany();
-        await Manager.Entity.deleteMany();
-        await mongoose.disconnect();
-    });
-
-    it('should return auth requirements', async () => {
-        const requirements = await manager.getAuthorizationRequirements();
-        expect(requirements).exists;
-        expect(requirements.type).toEqual('oauth2');
-    });
-});
